test: refactor FitPlugin into its own file and add coverage
