+++ src/fitSync.realFit.test.ts
@@ -13,7 +13,8 @@ import { Fit } from './fit';
 import { Vault } from 'obsidian';
 import { FakeLocalVault, FakeRemoteVault } from './testUtils';
 import { LocalVault } from './localVault';
-import { FitSettings, LocalStores } from '@main';
+import { FitSettings } from '@/fitSettings';
+import { LocalStores } from '@/localStores';
 import { VaultError } from './vault';
 import { fitLogger } from './logger';
 import { FileContent } from './util/contentEncoding';
